requirements.txt
httpx>=0.27
selectolax>=0.3
duckdb>=0.10
pydantic>=2.6
pyyaml>=6.0
croniter>=1.4
python-dotenv>=1.0
